Moved behaviour that affects multiple VideoPlayers into a separate media module. Refactored VideoPlayer's controls into a standalone component, called VideoControls.
