+++ src/app/media.js
@@ -1,0 +1,34 @@
+const { enqueue, subscribe } = require('./scheduler');
+const { proximityCheck } = require('./utils/misc');
+
+const AMBIENT_PLAYABLE_RANGE = 0.5;
+
+const players = [];
+
+const forEachPlayer = (module.exports.forEachPlayer = fn => players.forEach(fn));
+
+module.exports.registerPlayer = player => players.push(player);
+
+subscribe(function _checkIfPlayersNeedToBeToggled(client) {
+  forEachPlayer(player => {
+    if (player.isUserInControl || (!player.isAmbient && !player.isScrollplay)) {
+      return;
+    }
+
+    const rect = player.getRect();
+    const isInPlayableRange = player.isAmbient
+      ? proximityCheck(rect, client, AMBIENT_PLAYABLE_RANGE)
+      : proximityCheck(rect, client, (player.scrollplayPct || 0) / -100);
+
+    if (
+      (typeof player.isInPlayableRange === 'undefined' && isInPlayableRange) ||
+      (typeof player.isInPlayableRange !== 'undefined' && isInPlayableRange !== player.isInPlayableRange)
+    ) {
+      enqueue(function _toggleVideoPlay() {
+        player.togglePlayback(null, true);
+      });
+    }
+
+    player.isInPlayableRange = isInPlayableRange;
+  });
+});
